fix
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@1342 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ pde/src/common/BoundaryConditionsContainer.hpp
@@ -314,11 +314,11 @@ public:
         
             while (dirichIterator != mpDirichletMap[0]->end() )
             {
-                long node_index = dirichIterator->first->GetIndex();
+                int node_index = dirichIterator->first->GetIndex();
             
                 for (int col=0; col<cols; col++)
                 {
-                    value = (col == (NUM_UNKNOWNS*node_index + index_of_unknown)) ? 1.0 : 0.0;
+                    value = (col == (int)(NUM_UNKNOWNS*node_index + index_of_unknown)) ? 1.0 : 0.0;
                     MatSetValue(jacobian, node_index, col, value, INSERT_VALUES);
                 }
                 dirichIterator++;
